- Silence compiler on ORTE_TRANSPORT_KEY_FMT, it is fixed to llx  - No functional changes, just indentation and corrections to error    output.
This commit was SVN r12734.

+++ orte/util/pre_condition_transports.c
@@ -100,8 +100,8 @@ int orte_pre_condition_transports(orte_app_context_t **app_context, size_t num_c
     }
 #endif  /* !defined(__WINDOWS__) */
 
-    sprintf(string_key, ORTE_TRANSPORT_KEY_FMT, (long unsigned)unique_key[0], 
-            (long unsigned)unique_key[1]);
+    sprintf(string_key, ORTE_TRANSPORT_KEY_FMT, (long long unsigned)unique_key[0], 
+            (long long unsigned)unique_key[1]);
     string_key[sizeof string_key - 1] = '\0';
     
     if (NULL == (cs_env = mca_base_param_environ_variable("orte_precondition_transports",NULL,NULL))) {
